refactor: Type 폴더에 타입정리 이동
