+++ components/Main/location/locationListItem.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@chakra-ui/react';
 
-import { ILocationListItem } from '../types';
+import { ILocationListItem } from '@/types/Main';
 
 export default function LocationListItem({
   name,
